[BUGFIX] Do not query processed files which are not handled by the driver
Before trying to update S3 meta data, we need to ensure, the the
processed file is actually on an S3 storage.

We also need to make sure to resolve the file info with the driver
attached to the processed file storage, not the one passed, which is the
one from the original file, which could be a different one.