[FEATURE] Ensure usage of local file for hashing content
As a small mitigation for the huge amount of time it
takes to generate the hash of file content, we check
whether the file has just been uploaded in the same
request and use the local file to generate the hash from
in this case.

This avoids a re-download of the file, that has just been uploaded.