[BUGFIX] Flush file/folder existence runtime cache on change operations
Every time we flush the cache for a folder or file, it means we
deleted/renamed/added something to the storage. In this case
we need to make sure to also flush the runtime cache for file/folder
existence to make sure following operations yield correct results.

As drive by change we remove unnecessary calls to rtrim on
an already cleaned up file/folder identifier