[BUGFIX] Disable remote cache control feature
There is a considerable impact with adding cache control
headers to remote objects and that is that the modification time
will change of that object.

When performing this operation on files when meta data records are
changed within TYPO3, this leads to an update of the remote file,
directly invalidating all processed files, which are bound to the modification
date of the original file.

While trying to fix the causing errors, it was decided to better deactivate
this feature to avoid causing such issues