[FEATURE] Improve hashing code
Instead of directly operating on the stream with a file
function, we fetch the file for local processing,
to benefit from the local file cache functionality implemented,
which avoids downloading a file more than once during a request.

Additionally, we use a head request to try to extract the
md5 from the ETag and only fall back to file download if
this fails.